+++ .storybook/main.ts
@@ -1,25 +1,12 @@
-//
-// This source file is part of the Stanford Biodesign Digital Health ENGAGE-HF open-source project
-//
-// SPDX-FileCopyrightText: 2023 Stanford University and the project authors (see CONTRIBUTORS.md)
-//
-// SPDX-License-Identifier: MIT
-//
-
 import type { StorybookConfig } from "@storybook/react-vite";
 
 const config: StorybookConfig = {
-  stories: ["../**/*.mdx", "../**/*.stories.@(js|jsx|mjs|ts|tsx)"],
-  addons: [
-    "@storybook/addon-onboarding",
-    "@storybook/addon-essentials",
-    "@storybook/addon-interactions",
-  ],
+  stories: ["../stories/**/*.stories.@(js|jsx|mjs|ts|tsx)"],
+  addons: [],
   framework: {
     name: "@storybook/react-vite",
     options: {},
   },
-  docs: {},
   typescript: {
     reactDocgen: "react-docgen-typescript",
   },
