--- .storybook/preview.tsx
@@ -1,25 +1,0 @@
-//
-// This source file is part of the Stanford Biodesign Digital Health Spezi Web Design System open-source project
-//
-// SPDX-FileCopyrightText: 2024 Stanford University and the project authors (see CONTRIBUTORS.md)
-//
-// SPDX-License-Identifier: MIT
-//
-
-import { type Preview } from "@storybook/react";
-import "../modules/globals.css";
-
-const preview: Preview = {
-  parameters: {
-    layout: "centered",
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-  },
-  tags: ["autodocs"],
-};
-
-export default preview;
