+++ package.json
@@ -47,13 +47,7 @@
   },
   "devDependencies": {
     "@stanfordspezi/spezi-web-configurations": "^0.6.0",
-    "@storybook/addon-essentials": "^8.6.14",
-    "@storybook/addon-interactions": "^8.6.14",
-    "@storybook/addon-onboarding": "^8.6.14",
-    "@storybook/blocks": "^8.6.14",
-    "@storybook/react": "^8.6.14",
-    "@storybook/react-vite": "^8.6.14",
-    "@storybook/test": "^8.6.14",
+    "@storybook/react-vite": "^9.1.8",
     "@tailwindcss/postcss": "^4.1.13",
     "@tanstack/router-plugin": "^1.132.23",
     "@testing-library/jest-dom": "^6",
@@ -70,7 +64,7 @@
     "jsdom": "^27.0.0",
     "postcss": "^8",
     "prettier": "^3",
-    "storybook": "^8.6.14",
+    "storybook": "^9.1.8",
     "tailwindcss": "^4",
     "tailwindcss-animate": "^1.0.7",
     "typescript": "^5",
